src/auto_snippets/requirements.txt
functions-framework>=3.10.0
google-cloud-firestore>=2.22.0
google-cloud-secret-manager>=2.16.4
google-cloud-storage>=3.0.0
google-cloud-aiplatform>=1.130.0
requests>=2.32.3
beautifulsoup4>=4.12.0
PyYAML>=6.0.1
